·hasta aca andan los test de todo si suben algo primero vean que anden todos los testgit add . git add . git add .

+++ src/main/java/ar/edu/utn/frba/dds/models/entities/ubicacionGeografica/Coordenada.java
@@ -12,7 +12,7 @@ import javax.persistence.*;
 public class Coordenada {
     @Id
     @GeneratedValue( strategy = GenerationType.IDENTITY)
-    private Integer id_Coordenada;
+    private Long id_Coordenada;
     @Column(name = "latitud",columnDefinition = "DOUBLE")
     private Double latitud;
     @Column(name = "longitud",columnDefinition = "DOUBLE")
